requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
yfinance>=0.2.31
ta>=0.11.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
fpdf2>=2.7.6
bcrypt>=4.1.0
requests>=2.31.0
pytz>=2023.3
aiohttp>=3.9.0
pytest>=7.0.0
